backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.50.0
mem0ai>=0.1.0
tavily-python>=0.5.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
sse-starlette>=2.1.0
pydantic>=2.0.0
python-multipart>=0.0.9
pydantic-settings>=2.0.0
